refactor: コンポーネントフォルダ内を6ドメインに分割
コンポーネントの数が多くファイルを見つけづらくなったので分割しました。
