+++ app/components/auth/LoginForm.tsx
@@ -5,7 +5,7 @@ import { login } from "@/app/actions";
 import Link from "next/link";
 import { buttonConfig, headingConfig, textConfig } from "@/app/constants/ui";
 import { userFormConfig } from "@/app/constants/form";
-import LoadingOverlay from "@/app/components/LoadingOverlay";
+import LoadingOverlay from "@/app/components/ui/LoadingOverlay";
 
 const LoginForm = () => {
   const [state, formAction, pending] = useActionState(login, null);
