+++ app/components/expense/EditExpenseButton.tsx
@@ -6,8 +6,8 @@ import { updateExpense } from "@/app/actions";
 import { validationConfig } from "@/app/constants/validation";
 import { CATEGORIES, expenseFormConfig } from "@/app/constants/form";
 import { buttonConfig } from "@/app/constants/ui";
-import CancelButton from "@/app/components/CancelButton";
-import LoadingOverlay from "@/app/components/LoadingOverlay";
+import CancelButton from "@/app/components/ui/CancelButton";
+import LoadingOverlay from "@/app/components/ui/LoadingOverlay";
 
 export const EditExpenseButton = ({
   expenseId,
