+++ app/components/participant/EditParticipantButton.tsx
@@ -6,8 +6,8 @@ import { updateParticipant } from "@/app/actions";
 import { validationConfig } from "@/app/constants/validation";
 import { participantFormConfig } from "@/app/constants/form";
 import { buttonConfig } from "@/app/constants/ui";
-import CancelButton from "@/app/components/CancelButton";
-import LoadingOverlay from "@/app/components/LoadingOverlay";
+import CancelButton from "@/app/components/ui/CancelButton";
+import LoadingOverlay from "@/app/components/ui/LoadingOverlay";
 
 const EMAIL_PATTERN = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
