+++ app/components/spot/DeleteSpotButton.tsx
@@ -2,19 +2,25 @@
 
 import { useRouter } from "next/navigation";
 import { startTransition, useActionState } from "react";
-import { deleteTrip } from "@/app/actions";
+import { deleteSpot } from "@/app/actions";
 import { buttonConfig, confirmConfig } from "@/app/constants/ui";
+import { validationConfig } from "@/app/constants/validation";
 import { SubmitState } from "@/lib/types";
-import { validationConfig } from "../constants/validation";
-import LoadingOverlay from "@/app/components/LoadingOverlay";
+import LoadingOverlay from "@/app/components/ui/LoadingOverlay";
 
-const TripDelete = ({ tripId }: { tripId: string }) => {
+export const DeleteSpotButton = ({
+  spotId,
+  tripId,
+}: {
+  spotId: number;
+  tripId: string;
+}) => {
   const router = useRouter();
   const [state, action, pending] = useActionState<SubmitState>(
     async (_prevState) => {
       try {
-        await deleteTrip(tripId);
-        router.push("/");
+        await deleteSpot(spotId, tripId);
+        router.refresh();
         return { error: null };
       } catch (e) {
         return {
@@ -26,7 +32,7 @@ const TripDelete = ({ tripId }: { tripId: string }) => {
   );
 
   const handleDelete = () => {
-    if (!confirm(confirmConfig.deleteTrip)) return;
+    if (!confirm(confirmConfig.deleteSpot)) return;
     startTransition(action);
   };
 
@@ -35,14 +41,12 @@ const TripDelete = ({ tripId }: { tripId: string }) => {
       <button
         onClick={handleDelete}
         disabled={pending}
-        className="text-sm mr-4 px-3 py-1.5 rounded-lg border border-red-200 text-red-600 hover:bg-red-50 disabled:opacity-50 transition-colors duration-300 cursor-pointer"
+        className="ml-auto px-2 text-xs text-gray-400 hover:text-red-500 disabled:opacity-50 cursor-pointer"
       >
-        {buttonConfig.tripDelete}
+        {buttonConfig.delete}
       </button>
       {state.error && <p className="text-xs text-red-500">{state.error}</p>}
       <LoadingOverlay visible={pending} />
     </>
   );
 };
-
-export default TripDelete;
