+++ app/components/trip/TripAction.tsx
@@ -6,7 +6,7 @@ import { updateTripStatus } from "@/app/actions";
 import type { TripStatus } from "@/lib/types";
 import { buttonConfig } from "@/app/constants/ui";
 import { validationConfig } from "@/app/constants/validation";
-import LoadingOverlay from "@/app/components/LoadingOverlay";
+import LoadingOverlay from "@/app/components/ui/LoadingOverlay";
 
 const TripAction = ({
   tripId,
