+++ app/components/trip/TripCard.tsx
