+++ app/components/trip/TripDelete.tsx
@@ -2,25 +2,19 @@
 
 import { useRouter } from "next/navigation";
 import { startTransition, useActionState } from "react";
-import { deleteExpense } from "@/app/actions";
+import { deleteTrip } from "@/app/actions";
 import { buttonConfig, confirmConfig } from "@/app/constants/ui";
-import { validationConfig } from "@/app/constants/validation";
 import { SubmitState } from "@/lib/types";
-import LoadingOverlay from "@/app/components/LoadingOverlay";
+import { validationConfig } from "@/app/constants/validation";
+import LoadingOverlay from "@/app/components/ui/LoadingOverlay";
 
-export const DeleteExpenseButton = ({
-  expenseId,
-  tripId,
-}: {
-  expenseId: number;
-  tripId: string;
-}) => {
+const TripDelete = ({ tripId }: { tripId: string }) => {
   const router = useRouter();
   const [state, action, pending] = useActionState<SubmitState>(
     async (_prevState) => {
       try {
-        await deleteExpense(expenseId, tripId);
-        router.refresh();
+        await deleteTrip(tripId);
+        router.push("/");
         return { error: null };
       } catch (e) {
         return {
@@ -31,8 +25,8 @@ export const DeleteExpenseButton = ({
     { error: null },
   );
 
-  const handleDelete = async () => {
-    if (!confirm(confirmConfig.deleteExpense)) return;
+  const handleDelete = () => {
+    if (!confirm(confirmConfig.deleteTrip)) return;
     startTransition(action);
   };
 
@@ -41,12 +35,14 @@ export const DeleteExpenseButton = ({
       <button
         onClick={handleDelete}
         disabled={pending}
-        className="px-2 text-xs text-gray-400 hover:text-red-500 disabled:opacity-50 cursor-pointer"
+        className="text-sm mr-4 px-3 py-1.5 rounded-lg border border-red-200 text-red-600 hover:bg-red-50 disabled:opacity-50 transition-colors duration-300 cursor-pointer"
       >
-        {buttonConfig.delete}
+        {buttonConfig.tripDelete}
       </button>
       {state.error && <p className="text-xs text-red-500">{state.error}</p>}
       <LoadingOverlay visible={pending} />
     </>
   );
 };
+
+export default TripDelete;
