+++ app/components/ui/AchievementRateBar.tsx
